Created a simple node.js server for the app
Created a new file server.js, which is a node.js file that serves the
web application on port 3000 of localhost and displays a 404 error on
attempt to access any other file. Added a package-json and installed
MongoDB as a dependency. Moved .css and front-end .js files into folders
called css and js respectively.